Remove Category bottom and added in the main page :iphone:
