+++ app/src/main/res/layout/activity_main_page.xml
@@ -7,8 +7,6 @@
     android:fitsSystemWindows="true"
     tools:context=".activities.MainPageActivity">
 
-
-
     <FrameLayout
         android:id="@+id/container"
         android:layout_width="match_parent"
