+++ app/src/main/res/layout/cardview_category.xml
@@ -1,43 +1,45 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="200dp"
-    android:layout_marginEnd="50dp"
-    android:layout_marginStart="50dp"
-    android:layout_marginBottom="15dp"
-    android:layout_marginTop="10dp"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    app:cardCornerRadius="10dp">
+    android:layout_width="170dp"
+    android:layout_height="90dp"
+    android:layout_marginHorizontal="8dp"
+    app:cardBackgroundColor="@color/Dark_blue_light"
+    app:cardCornerRadius="16dp">
     
-    <FrameLayout
+    <LinearLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="match_parent"
+        android:gravity="center_vertical"
+        android:layout_margin="8dp"
+        android:orientation="vertical">
 
         <ImageView
             android:id="@+id/imageview_category"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
+            android:layout_width="5dp"
+            android:layout_height="5dp"
             android:src="@drawable/ic_image"
-            android:scaleType="centerCrop"/>
-
-        <View
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="@color/black"
-            android:alpha="0.5"/>
+            android:layout_gravity="center"
+            />
 
         <TextView
             android:id="@+id/textview_category"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:text="Categoria"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="categorias"
+            android:layout_gravity="center_horizontal"
             android:textColor="@color/white"
-            android:gravity="center"
             android:textStyle="bold"
-            android:textSize="20sp"/>
+            android:textSize="16sp"
+            />
 
-    </FrameLayout>
+        <View
+            android:id="@+id/linear"
+            android:layout_marginTop="4dp"
+            android:background="@color/Blue_light"
+            android:layout_width="match_parent"
+            android:layout_height="3dp"/>
 
-
+    </LinearLayout>
 
 </androidx.cardview.widget.CardView>
