+++ app/src/main/res/menu/bottom_navigation_home.xml
@@ -6,11 +6,6 @@
         android:title=""
         />
 
-    <item android:id="@+id/item_category"
-        android:icon="@drawable/outline_category_24"
-        android:title=""
-        />
-
     <item android:id="@+id/item_new_post"
         android:icon="@drawable/outline_add_box_24"
         android:title=""
