+++ app/src/main/res/values/colors.xml
@@ -12,4 +12,5 @@
     <color name="Dark_blue">#1B2A41</color>
     <color name="Dark_blue_light">#051B5A</color>
     <color name="Gray_light">#D9D9D9</color>
+    <color name="Blue_light">#2E86C1</color>
 </resources>
